+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Key: [+] added, [-] removed, [~] modified.
 
 ## 2.1.0
 + [+] Added NSFW module with an [__efukt__] command, forcing me to adopt Jsoup which is actually pretty good. 
++ [+] Added [__clearqueue__] command to easily clear the queue without having to remove and recall the bot.
 
 ## 2.0.0